chore: update merkle tree audit scope (#20666)
Removes dead code from native merkle_tree module and updates audit scope
doc accordingly
