+++ barretenberg/cpp/src/barretenberg/benchmark/CMakeLists.txt
@@ -7,7 +7,6 @@ add_subdirectory(chonk_bench)
 add_subdirectory(pippenger_bench)
 add_subdirectory(relations_bench)
 add_subdirectory(poseidon2_bench)
-add_subdirectory(merkle_tree_bench)
 add_subdirectory(indexed_tree_bench)
 add_subdirectory(append_only_tree_bench)
 add_subdirectory(ultra_bench)
