--- barretenberg/cpp/src/barretenberg/benchmark/merkle_tree_bench/CMakeLists.txt
@@ -1,1 +1,0 @@
-barretenberg_module(merkle_tree_bench crypto_poseidon2 ext_starknet_crypto_poseidon crypto_pedersen_hash crypto_merkle_tree)
