--- barretenberg/cpp/src/barretenberg/crypto/merkle_tree/index.hpp
@@ -1,12 +1,0 @@
-// === AUDIT STATUS ===
-// internal:    { status: Planned, auditors: [Raju], commit: }
-// external_1:  { status: not started, auditors: [], commit: }
-// external_2:  { status: not started, auditors: [], commit: }
-// =====================
-
-#pragma once
-#include "hash.hpp"
-#include "hash_path.hpp"
-#include "memory_store.hpp"
-#include "memory_tree.hpp"
-#include "merkle_tree.hpp"
